feat: 중복 닉네임 검증 api 추가 (#237)
* refactor: SiteUser에서 MyPage로 서비스명 변경하여 역할 명확화

* feat: 닉네임 중복 검증 api 추가

* style: 컨트롤러 메서드 인자 개행 컨벤션 적용

* refactor: checkNicknameExists로 함수명 개선
