+++ src/main/java/com/example/solidconnection/siteuser/dto/NicknameExistsResponse.java
@@ -1,0 +1,9 @@
+package com.example.solidconnection.siteuser.dto;
+
+public record NicknameExistsResponse(
+        boolean exists
+) {
+    public static NicknameExistsResponse from(boolean exists) {
+        return new NicknameExistsResponse(exists);
+    }
+}
